requirements.txt
streamlit==1.27.0
transformers==4.35.0
torch==2.0.1
PyPDF2==3.0.0
python-docx==0.8.11
